requirements.txt
# Core dependencies (pinned for stability)
streamlit>=1.28.0,<1.30.0
pandas>=2.0.0,<2.1.0
numpy>=1.24.0,<1.25.0
plotly>=5.15.0,<5.16.0
python-dotenv>=1.0.0,<1.1.0

# PDF processing
pdf2image>=1.16.0,<1.17.0
pillow>=10.0.0,<10.1.0

# Database - let pip pick compatible version
supabase>=1.0.0

# Ollama client - let pip pick compatible version
ollama>=0.1.0

# PostgreSQL driver
psycopg2-binary>=2.9.0

# For fetching ollama models
requests>=2.31.0

# For Mistral API
mistralai>=0.1.5